+++ @angular-generic-table/core/pipes/gt-order-by.pipe.ts
@@ -9,7 +9,7 @@ export class GtOrderByPipe<R extends GtRow> implements PipeTransform {
 
   /** Return property */
   private getProperty = function(array: Array<any>, key: string){
-    for (let i = 0; i < array.length; i++){
+    for (let i = 0; i < array.length; i++) {
       if (array[i].objectKey === key) {
         return array[i];
       }
@@ -17,14 +17,14 @@ export class GtOrderByPipe<R extends GtRow> implements PipeTransform {
   };
 
   /** Return sort function */
-  private getSortFunction(field: any){
-    if (!field){
-      console.log('error trying to sort undefined field');
+  private getSortFunction(field: any) {
+    if (!field) {
+      // console.log('error trying to sort undefined field');
       return false;
     }
-    if (typeof field.sort === 'function'){
+    if (typeof field.sort === 'function') {
       return field.sort;
-    } else if (typeof field.value === 'function'){
+    } else if (typeof field.value === 'function') {
       return field.value;
     } else {
       return false;
@@ -32,81 +32,77 @@ export class GtOrderByPipe<R extends GtRow> implements PipeTransform {
   }
 
 
-  static _orderByComparator(a: any, b: any): number{
+  static _orderByComparator(a: any, b: any): number {
 
     // sort boolean values as strings
     if (typeof a === 'boolean') {a = a.toString(); }
     if (typeof b === 'boolean') {b = b.toString(); }
 
-    if ((isNaN(parseFloat(a)) || !isFinite(a)) || (isNaN(parseFloat(b)) || !isFinite(b))){
+    if ((isNaN(parseFloat(a)) || !isFinite(a)) || (isNaN(parseFloat(b)) || !isFinite(b))) {
 
-      if (b === null || typeof b === 'undefined' && (a !== null && typeof a !== 'undefined' )) return 1;
-      if (a === null || typeof a === 'undefined' && (b !== null && typeof b !== 'undefined' )) return -1;
+      if (b === null || typeof b === 'undefined' && (a !== null && typeof a !== 'undefined' )) { return 1; }
+      if (a === null || typeof a === 'undefined' && (b !== null && typeof b !== 'undefined' )) { return -1; }
 
-      //Isn't a number so lowercase the string to properly compare
+      // isn't a number so lowercase the string to properly compare
       try {
-        if (a.toLowerCase() < b.toLowerCase()) return -1;
-        if (a.toLowerCase() > b.toLowerCase()) return 1;
+        if (a.toLowerCase() < b.toLowerCase()) { return -1; }
+        if (a.toLowerCase() > b.toLowerCase()) { return 1; }
       } catch (error) {
         return 0;
       }
-    }
-    else{
-      //Parse strings as numbers to compare properly
-      if (parseFloat(a) < parseFloat(b)) return -1;
-      if (parseFloat(a) > parseFloat(b)) return 1;
+    } else {
+      // parse strings as numbers to compare properly
+      if (parseFloat(a) < parseFloat(b)) { return -1; }
+      if (parseFloat(a) > parseFloat(b)) { return 1; }
     }
 
-    return 0; //equal each other
+    return 0; // equal each other
   }
 
   transform(input: any, sortByProperties: Array<string>, fields: GtConfigField<R, any>[], refreshSorting: boolean, refreshData: number): any {
 
-    if (!Array.isArray(input) || input === null) return input;
-    if (!Array.isArray(sortByProperties) || (Array.isArray(sortByProperties) && sortByProperties.length == 1)){
+    if (!Array.isArray(input) || input === null) { return input; }
+    if (!Array.isArray(sortByProperties) || (Array.isArray(sortByProperties) && sortByProperties.length === 1)) {
 
       const propertyToCheck: string = sortByProperties[0];
-      const desc = propertyToCheck.substr(0, 1) == '-';
+      const desc = propertyToCheck.substr(0, 1) === '-';
 
-      //Basic array
-      if (!propertyToCheck || propertyToCheck == '-' || propertyToCheck == '+'){
+      // basic array
+      if (!propertyToCheck || propertyToCheck === '-' || propertyToCheck === '+') {
         return !desc ? input.sort() : input.sort().reverse();
-      }
-      else {
-        const property: string = propertyToCheck.substr(0, 1) == '+' || propertyToCheck.substr(0, 1) == '-'
+      } else {
+        const property: string = propertyToCheck.substr(0, 1) === '+' || propertyToCheck.substr(0, 1) === '-'
           ? propertyToCheck.substr(1)
           : propertyToCheck;
 
         // check if custom sort function is defined
         const sortFunction: any = this.getSortFunction(this.getProperty(fields, property));
 
-        return input.sort(function(a: any, b: any){
+        return input.sort(function(a: any, b: any) {
 
           // use custom sort function if one is defined
           const propertyA = sortFunction === false ? a[property] : sortFunction(a);
           const propertyB = sortFunction === false ? b[property] : sortFunction(b);
 
           // if both values are undefined...
-          if (typeof propertyA === 'undefined' && typeof propertyB === 'undefined' ){
+          if (typeof propertyA === 'undefined' && typeof propertyB === 'undefined' ) {
             // ...skip comparison
             return;
           }
-
           return !desc ? GtOrderByPipe._orderByComparator(propertyA, propertyB) : -GtOrderByPipe._orderByComparator(propertyA, propertyB);
         });
       }
-    }
-    else {
-      //Loop over property of the array in order and sort
+    } else {
+      // loop over property of the array in order and sort
       return input.sort((a: any, b: any) => {
-        //console.log('multiple');
-        for (let i = 0; i < sortByProperties.length; i++){
-          const desc = sortByProperties[i].substr(0, 1) == '-';
-          const property = sortByProperties[i].substr(0, 1) == '+' || sortByProperties[i].substr(0, 1) == '-'
+        // console.log('multiple');
+        for (let i = 0; i < sortByProperties.length; i++) {
+          const desc = sortByProperties[i].substr(0, 1) === '-';
+          const property = sortByProperties[i].substr(0, 1) === '+' || sortByProperties[i].substr(0, 1) === '-'
             ? sortByProperties[i].substr(1)
             : sortByProperties[i];
 
-          //console.log(property);
+          // console.log(property);
           // check if custom sort function is defined
           const sortFunction: any = this.getSortFunction(this.getProperty(fields, property));
 
@@ -116,11 +112,11 @@ export class GtOrderByPipe<R extends GtRow> implements PipeTransform {
 
           const comparison = !desc ? GtOrderByPipe._orderByComparator(propertyA, propertyB) : -GtOrderByPipe._orderByComparator(propertyA, propertyB);
 
-          //Don't return 0 yet in case of needing to sort by next property
-          if (comparison != 0) return comparison;
+          // don't return 0 yet in case of needing to sort by next property
+          if (comparison !== 0) { return comparison; }
         }
 
-        return 0; //equal each other
+        return 0; // equal each other
       });
     }
   }